Move upgrade functions to ShaderGraph.Upgrades.cs
