+++ game/addons/tools/Code/ShaderGraph/ShaderGraph.Upgrades.cs
@@ -1,0 +1,102 @@
+﻿using Editor.NodeEditor;
+using System.Reflection;
+using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
+
+namespace Editor.ShaderGraph;
+
+partial class ShaderGraph
+{
+	/// <summary>
+	/// Check if a legacy parameter node should be upgraded to SubgraphInput.
+	/// </summary>
+	private static bool ShouldUpgradeToSubgraphInput( string typeName, JsonElement element )
+	{
+		// Only upgrade if it's a parameter node type
+		if ( !IsParameterNodeType( typeName ) )
+			return false;
+
+		// Only upgrade if it has a name (indicating it's meant to be an input)
+		if ( element.TryGetProperty( "Name", out var nameProperty ) )
+		{
+			var name = nameProperty.GetString();
+			return !string.IsNullOrWhiteSpace( name );
+		}
+
+		return false;
+	}
+
+	/// <summary>
+	/// Check if the type name represents a parameter node
+	/// </summary>
+	private static bool IsParameterNodeType( string typeName )
+	{
+		return typeName switch
+		{
+			"Float" => true,
+			"Float2" => true,
+			"Float3" => true,
+			"Float4" => true,
+			"TextureSampler" => true,
+			_ => false
+		};
+	}
+
+	/// <summary>
+	/// Create a new SubgraphInput node from a legacy parameter node
+	/// </summary>
+	private SubgraphInput CreateUpgradedSubgraphInput( string typeName, JsonElement element, JsonSerializerOptions options )
+	{
+		var subgraphInput = new SubgraphInput();
+
+		// Copy basic node properties
+		DeserializeObject( subgraphInput, element, options );
+
+		// Set input name from the parameter's Name property
+		if ( element.TryGetProperty( "Name", out var nameProperty ) )
+		{
+			subgraphInput.InputName = nameProperty.GetString();
+		}
+
+		// Map the parameter type to InputType and set default values
+		switch ( typeName )
+		{
+			case "Float":
+				subgraphInput.InputType = InputType.Float;
+				if ( element.TryGetProperty( "Value", out var floatValue ) )
+				{
+					subgraphInput.DefaultFloat = floatValue.GetSingle();
+				}
+				break;
+
+			case "Float2":
+				subgraphInput.InputType = InputType.Float2;
+				if ( element.TryGetProperty( "Value", out var float2Value ) )
+				{
+					var vector2 = JsonSerializer.Deserialize<Vector2>( float2Value.GetRawText(), options );
+					subgraphInput.DefaultFloat2 = vector2;
+				}
+				break;
+
+			case "Float3":
+				subgraphInput.InputType = InputType.Float3;
+				if ( element.TryGetProperty( "Value", out var float3Value ) )
+				{
+					var vector3 = JsonSerializer.Deserialize<Vector3>( float3Value.GetRawText(), options );
+					subgraphInput.DefaultFloat3 = vector3;
+				}
+				break;
+
+			case "Float4":
+				subgraphInput.InputType = InputType.Color;
+				if ( element.TryGetProperty( "Value", out var float4Value ) )
+				{
+					var color = JsonSerializer.Deserialize<Color>( float4Value.GetRawText(), options );
+					subgraphInput.DefaultColor = color;
+				}
+				break;
+		}
+
+		return subgraphInput;
+	}
+}
